Dedicated parsing into an AST index.
Creating an AST is significantly faster than the equivalent
`jsonv::parse` call (approximately 10x speedup). This completely removes
the older `tokenizer` structure and replaces it was a flat `parse_index`
structure.

This commit introduces #145, which calls to revisit `parse_options` and
the associated `parse_error`. Now that the parse phase and extraction to
a JSON `value` are separate, the old configuration options make less
sense. This also enables work on #150, which calls to extract directly
from source text.

- Fixes #147
- Fixes #96 through `ast_node::integer` and `ast_node::decimal`
- Closes #97 by removal of `tokenizer`